+++ tasks/ps-rule-installV1/powershell.ps1
@@ -22,6 +22,8 @@ param (
     [System.Boolean]$PreRelease = (Get-VstsInput -Name 'prerelease' -AsBool)
 )
 
+Write-Host "`#`#vso[task.logissue type=warning]This version of the task is deprecated. Please upgrade to the latest version by using 'ps-rule-install@2'. See https://aka.ms/ps-rule-pipelines/upgrade for details.";
+
 if ($Env:SYSTEM_DEBUG -eq 'true') {
     $VerbosePreference = 'Continue';
 }