+++ CHANGELOG.md
@@ -22,6 +22,8 @@ What's changed since v2.7.0:
     [#694](https://github.com/microsoft/PSRule-pipelines/issues/694)
     - Deprecated task versions `@0` and `@1` will be removed from _v3.0.0_.
     - Upgrade to `@2` task versions to ensure your pipelines continue to work.
+  - Updated Node 10 tasks to require minimum agent version by @BernieWhite.
+    [#692](https://github.com/microsoft/PSRule-pipelines/issues/692)
 - Engineering:
   - Bump azure-pipelines-task-lib to v4.2.0.
     [#660](https://github.com/microsoft/PSRule-pipelines/pull/660)